refactor: extract rates loading into useRates hook
