+++ Frontend/Task/app/Main.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { PairsSelector } from './PairsSelector';
 import { RateList } from './RateList';
-import { fetchRates, Rates } from './dataFetching/fetchRates';
 import { compareRates } from './compareRates';
 import { useConfig } from './useConfig';
+import { useRates } from './useRates';
 
 type Props = {
   configUrl: string;
@@ -31,39 +31,11 @@ export const Main = ({
     });
   }, []);
 
-  const [[currentRates, previousRates], setRates] = useState<[Rates, Rates]>([
-    {},
-    {},
-  ]);
-
-  const loadRates = useCallback(async () => {
-    if (config) {
-      const freshestRates = await fetchRates(
-        ratesUrl,
-        Object.keys(config.currencyPairs),
-      );
-      setRates(([formerCurrentRates, formerPreviousRates]) => [
-        {
-          ...formerCurrentRates,
-          ...freshestRates,
-        },
-        {
-          ...formerPreviousRates,
-          ...formerCurrentRates,
-        },
-      ]);
-    }
-  }, [config, ratesUrl]);
-
-  useEffect(() => {
-    if (config) {
-      loadRates();
-      const intervalId = setInterval(loadRates, ratesRefreshMilliseconds);
-      return () => {
-        clearInterval(intervalId);
-      };
-    }
-  }, [config, loadRates, ratesRefreshMilliseconds]);
+  const [currentRates, previousRates] = useRates(
+    config,
+    ratesUrl,
+    ratesRefreshMilliseconds,
+  );
 
   if (config) {
     const pairs = Object.entries(config.currencyPairs).map(([id, pair]) => ({
